Core #23

+++ src/main/java/ru/core/l1/figure/Rect.java
@@ -1,19 +1,19 @@
 package ru.core.l1.figure;
 
 public class Rect {
-    private final float width;
-    private final float length;
+    private final double width;
+    private final double length;
 
-    public Rect(float width, float length) {
+    public Rect(double width, double length) {
         this.width = width;
         this.length = length;
     }
 
-    public float perimeter() {
+    public double perimeter() {
         return 2 * (width + length);
     }
 
-    public float squareRect() {
+    public double squareRect() {
         return width * length;
     }
 }
